feat: add fixed positioning for tooltip
